backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
pandas
numpy
joblib
websockets
python-multipart
pydantic
pydantic-settings
python-dotenv
httpx
